小更新:将一言插件的插件名改为hitokoto
如上.
